+++ jest.config.js
@@ -2,5 +2,5 @@ module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
   verbose: true,
-  testMatch: ["**/test/*.test.ts"],
+  testMatch: ["**/test/services/*.test.ts"],
 };
